refactored create/delete portfolio honors
