+++ src/components/createOrDeletePortfolio/CreatePortfolioHook.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { useCookies } from 'react-cookie';
 import {toast} from "react-toastify";
 import { useHistory } from "react-router-dom";
-import {portfolioUrl} from "../api/api";
+import {portfolioUrl} from "../../api/api";
 
 const useForm = (initialValues: any, portfolioValidate: any) => {
     const [inputs, setInputs] = useState(initialValues)
